url shortener. After review refactoring Clean up project root folder
