+++ src/main/java/faang/school/urlshortenerservice/hash/Base62Encoder.java
@@ -1,4 +1,4 @@
-package faang.school.urlshortenerservice;
+package faang.school.urlshortenerservice.hash;
 
 import faang.school.urlshortenerservice.config.HashEncoderProperties;
 import org.springframework.stereotype.Component;
